==> examples/beetles/2-mountain-pine-model.Rmd <==
# Robust spatiotemporal mountain pine beetle example

```{r, echo = FALSE, cache=FALSE, message=FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>",
  fig.height=8.5,
  fig.width=9
)
knitr::opts_knit$set(root.dir = "../../")
library(dplyr)
library(ggplot2)
library(glmmfields)
library(rstan)
library(viridis)
library(assertthat)
rstan_options(auto_write = TRUE)
options(mc.cores = parallel::detectCores())
# devtools::install_github("seananderson/ggsidekick")
theme_set(ggsidekick::theme_sleek())
```

Read in and set up the data.

```{r, echo=FALSE}
library(mapdata)
library(maps)
library(rgdal)
mpc <- ggplot2::map_data("worldHires", "Canada")
mps <- ggplot2::map_data("state")

mpc$group <- mpc$group + max(mps$group)
mp <- rbind(mpc,mps)

mp2 <- mp
coordinates(mp2) <- c("long", "lat")
proj <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0"
proj4string(mp2) <- CRS("+proj=longlat +datum=NAD83")
proj4string(mp2)

mp2 <- spTransform(mp2, CRS(proj))
proj4string(mp2)
```

```{r preparation, cache=FALSE, eval=TRUE}
id <- "mountain-pine-beetle-pnw-raster"
nbin <- 500

d <- readRDS(paste0("examples/beetles/",
  id, "-", "dataframe", "-", nbin, "x", nbin, ".rds"))

d_highres <- readRDS(paste0("examples/beetles/",
  id, "-", "dataframe", "-", nbin, "x", nbin, "highres.rds"))

# rescale
d$x <- d$x/1e5
d$y <- d$y/1e5

d <- dplyr::filter(d, cover > 0)
d_highres <- dplyr::filter(d_highres, cover > 0)

ggplot(d_highres, aes(x, y)) +
  geom_point(size = 0.1, alpha = 0.1) +
  facet_wrap(~year)

ggplot(d, aes(x, y, fill = log(cover))) +
  geom_tile() +
  facet_wrap(~year) +
  scale_fill_viridis(option = "D") +
  coord_fixed()

ggplot(d, aes(x, y, fill = cover)) + geom_tile() +
  facet_wrap(~year) +
  scale_fill_viridis(option = "D") +
  coord_fixed()

d <- mutate(d, row_number = seq_len(nrow(d)))
set.seed(1)
d <- group_by(d, year) %>%
  mutate(hold_out = row_number %in%
      sample(row_number, size = 25L)) %>%
  ungroup()

saveRDS(d, file = "examples/beetles/mountain-pine-beetle-data.rds")

# or the intercept prior will be off:
assert_that(mean(log(d$cover+0.5)) > -5 & mean(log(d$cover+0.5)) < 5)
# make sure data volume is reasonable:
assert_that(nrow(d) > 500)
assert_that(nrow(d) < 5000)
```

Fit the model. First the multivariate t model.

```{r model, cache=FALSE, warning=FALSE, message=FALSE, results='hide'}
model_file <- paste0("examples/beetles/", id, "-",
  "mvt-lognormal", "-", nbin, "x", nbin, ".rda")
if (!file.exists(model_file)) {
  mvt <- glmmfields(cover ~ 1,
    data = dplyr::filter(d, !hold_out),
    time = "year", lon = "x", lat = "y",
    nknots = 20,
    family = lognormal(link = "log"),
    prior_gp_sigma = half_t(3, 0, 1),
    prior_gp_theta = half_t(3, 0, 2),
    prior_intercept = student_t(1000, 0, 10),
    prior_rw_sigma = half_t(3, 0, 2),
    prior_sigma = half_t(3, 0, 3),
    prior_phi = half_t(1000, 0, 0.5),
    estimate_ar = TRUE,
    estimate_df = TRUE,
    year_re = TRUE,
    chains = 4, iter = 3000, thin = 1,
    control = list(adapt_delta = 0.95, max_treedepth = 20),
    save_log_lik = TRUE,
    df_lower_bound = 2)
  save(mvt, file = model_file)
} else {
  load(model_file)
}
# print(mvt)
# shinystan::launch_shinystan(mvt$model)
```

Now the multivariate normal model.

```{r, warning=FALSE, message=FALSE, results='hide'}
model_file <- paste0("examples/beetles/", id, "-",
  "mvn-lognormal", "-", nbin, "x", nbin, ".rda")
if (!file.exists(model_file)) {
  mvn <- glmmfields(cover ~ 1,
    data = dplyr::filter(d, !hold_out),
    time = "year", lon = "x", lat = "y",
    nknots = 20,
    family = lognormal(link = "log"),
    prior_gp_sigma = half_t(3, 0, 1),
    prior_gp_theta = half_t(3, 0, 2),
    prior_intercept = student_t(1000, 0, 10),
    prior_rw_sigma = half_t(3, 0, 2),
    prior_sigma = half_t(3, 0, 3),
    prior_phi = half_t(1000, 0, 0.5),
    estimate_ar = TRUE,
    estimate_df = FALSE,
    year_re = TRUE,
    fixed_df_value = 1000,
    chains = 4, iter = 3000, thin = 1,
    control = list(adapt_delta = 0.95, max_treedepth = 20),
    save_log_lik = TRUE)
  save(mvn, file = model_file)
} else {
  load(model_file)
}
# print(mvn)
# shinystan::launch_shinystan(mvn$model)
```

Let's make sure both models converged:

```{r}
bt <- broom::tidyMCMC(mvt$model, ess = TRUE, rhat = TRUE)
bn <- broom::tidyMCMC(mvn$model, ess = TRUE, rhat = TRUE)
max(bt$rhat)
min(bt$ess)
assertthat::assert_that(min(bt$ess) > 100)
assertthat::assert_that(max(bt$rhat) < 1.05)
max(bn$rhat)
min(bn$ess)
assertthat::assert_that(min(bn$ess) > 100)
assertthat::assert_that(max(bn$rhat) < 1.05)
```

Plot the predictions and check the model. Although functions for most of these plots are contained in the package, we will extract the data and make the plots ourselves so they can be more customized.

```{r plots, cache=FALSE, eval=TRUE}
pred <- predict(mvt, interval = "confidence", conf_level = 0.95,
  newdata = d)
pred <- pred %>% mutate(x = d$x, y = d$y, observed = d$cover,
  year = d$year, residual = log(observed) - estimate)

g <- ggplot(pred, aes(x, y, fill = estimate)) +
  geom_tile() + facet_wrap(~year) +
  scale_fill_viridis(option = "D") +
  coord_fixed()
print(g)

g <- ggplot(dplyr::filter(pred),
  aes(x, y, fill = exp(estimate))) +
  geom_tile() + facet_wrap(~year) +
  scale_fill_viridis(option = "D") +
  coord_fixed()
print(g)

g <- ggplot(pred, aes(x, y, fill = residual)) +
  geom_tile() + facet_wrap(~year) +
  scale_fill_gradient2() +
  coord_fixed()
print(g)

ggplot(pred, aes(estimate, residual)) +
  geom_point(alpha = 0.5) + facet_wrap(~year) +
  geom_hline(yintercept = 0, lty = 2) +
  geom_smooth(method = "loess", se = FALSE, colour = "red")

ggplot(pred, aes(log(observed), estimate)) +
  geom_point(alpha = 0.2) +
  geom_pointrange(aes(ymin = conf_low, ymax = conf_high), alpha = 0.1) +
  coord_fixed() +
  geom_abline(intercept = 0, slope = 1, colour = "red")
```

Let's compare the multivariate t and multivariate normal predictions.

```{r}
pred_mvn <- predict(mvn, interval = "confidence", conf_level = 0.95,
  newdata = d)
pred_mvn <- pred_mvn %>% mutate(x = d$x, y = d$y, observed = d$cover,
  year = d$year, residual = log(observed) - estimate)

g <- ggplot(dplyr::filter(pred_mvn),
  aes(x, y, fill = exp(estimate))) +
  geom_tile() + facet_wrap(~year) +
  scale_fill_viridis(option = "D") +
  coord_fixed()
print(g)

g <- ggplot(pred_mvn, aes(x, y, fill = residual)) +
  geom_tile() + facet_wrap(~year) +
  scale_fill_gradient2() +
  coord_fixed()
print(g)

pred2 <- pred
pred2$mvn_estimate <- pred_mvn$estimate
pred2$mvt_estimate <- pred$estimate
g <- ggplot(pred2, aes(mvt_estimate, mvn_estimate)) +
  geom_point() + facet_wrap(~year) +
  coord_fixed() +
  geom_abline(intercept = 0, slope = 1, colour = "red")
print(g)

g <- ggplot(pred2, aes(mvt_estimate, mvn_estimate)) +
  aes(x, y, fill = exp(mvt_estimate) - exp(mvn_estimate)) +
  geom_tile() + facet_wrap(~year) +
  scale_fill_gradient2() +
  coord_fixed()
print(g)

```

```{r}
combined <- bind_rows(
  mutate(pred, model = "mvt"),
  mutate(pred_mvn, model = "mvn"))

combined <- mutate(combined, id = paste(x, y))

set.seed(1234)
ids <- sample(unique(combined$id), size = 10)
jitter <- 0.2
dplyr::filter(combined, id %in% ids) %>%
  mutate(id_numeric = as.numeric(as.factor(id))) %>%
  mutate(id_numeric = ifelse(model == "mvn", id_numeric - jitter,
    id_numeric + jitter)) %>%
  ggplot(aes(x = id_numeric, y = estimate, ymin = conf_low,
    ymax = conf_high, color = model)) +
  geom_pointrange(size = 0.25) +
  facet_wrap(~year) +
  coord_flip() +
  geom_point(aes(x = round(id_numeric), y = log(observed)),
    color = "black", pch = 21) +
  xlab("y") + ylab("Point ID")
```

What about the holdout data?

```{r}
combined$hold_out <- c(d$hold_out, d$hold_out)

dplyr::filter(combined, hold_out) %>%
  ggplot(aes(log(observed), estimate)) +
  geom_point(alpha = 0.2) +
  geom_pointrange(aes(ymin = conf_low, ymax = conf_high), alpha = 0.1) +
  coord_fixed() +
  geom_abline(intercept = 0, slope = 1, colour = "red") +
  facet_wrap(~model)
```

Let's look at the mean squared error between the 2 models for the holdout data:

```{r}
mse <- dplyr::filter(combined, hold_out) %>%
  group_by(model, year) %>%
  summarize(mse = round(mean(residual^2), 6))
ggplot(mse, aes(year, mse, color = model)) +
  geom_line()

mse <- dplyr::filter(combined, hold_out) %>%
  group_by(model) %>%
  summarize(mse = round(mean(residual^2), 2))
print(mse)
```

What about the mean squared error for all of the data?

```{r}
mse <- dplyr::filter(combined) %>%
  group_by(model, year) %>%
  summarize(mse = round(mean(residual^2), 6))
ggplot(mse, aes(year, mse, color = model)) +
  geom_line()

mse <- dplyr::filter(combined) %>%
  group_by(model) %>%
  summarize(mse = round(mean(residual^2), 2))
print(mse)
```

Let's look at the observed and predicted values with confidence intervals for 10 of the holdout data points. In the simulations we've shown that the predictions in years with moderately large spikes can get inflated if we use the MVN model and the confidence intervals inflate to obtain the same correct coverage.

```{r}
jitter <- 0.2
dplyr::filter(combined, hold_out) %>%
  group_by(year, model) %>%
  mutate(id_numeric = 1:n()) %>%
  dplyr::filter(id_numeric <= 10) %>%
  mutate(id_numeric = ifelse(model == "mvn", id_numeric - jitter,
    id_numeric + jitter)) %>%
  ggplot(aes(x = id_numeric, y = estimate, ymin = conf_low,
    ymax = conf_high, color = model)) +
  geom_pointrange(size = 0.25) +
  facet_wrap(~year, scales = "free_x") +
  coord_flip() +
  geom_point(aes(x = round(id_numeric), y = log(observed)),
    color = "black", pch = 21) +
  xlab("y") + ylab("Point ID")
```

What about the width of the confidence intervals?

```{r}
cis <- mutate(combined, conf_width = conf_high - conf_low) %>%
  group_by(year, x, y) %>%
  summarize(conf_width_ratio = conf_width[model=="mvn"]/conf_width[model=="mvt"],
    hold_out = unique(hold_out)) %>%
  ungroup() %>%
  group_by(year) %>%
  mutate(median_ratio = median(conf_width_ratio))

dplyr::filter(cis) %>%
  ggplot(aes(conf_width_ratio)) +
  geom_histogram() +
  geom_vline(aes(xintercept = median(conf_width_ratio)))

dplyr::filter(cis) %>%
  ggplot(aes(conf_width_ratio)) +
  geom_histogram() +
  facet_wrap(~year) +
  geom_vline(xintercept = 1) +
  geom_vline(aes(xintercept = median_ratio), color = "red") +
  xlab("Ratio of prediction confidence intervals (MVN/MVT)")

median(cis$conf_width_ratio)
```

How are the model parameters themselves different?

```{r}
pars <- c("gp_sigma", "gp_theta", "sigma[1]")

bt <- broom::tidyMCMC(mvt$model, estimate.method = "median", conf.int = TRUE)
bn <- broom::tidyMCMC(mvn$model, estimate.method = "median", conf.int = TRUE)
bt$model <- "mvt"
bn$model <- "mvn"
b <- bind_rows(bt, bn)

dplyr::filter(b, term %in% pars) %>%
  ggplot(aes(term, estimate, ymin = conf.low,
    ymax = conf.high, color = model)) +
  geom_pointrange(position = position_dodge(width = 0.2)) +
  coord_flip()

bayesplot::mcmc_areas(as.matrix(mvt$model), pars = "df[1]")
```

Let's look at the coverage of the credible intervals between the 2 models:

```{r}
pred_mvt <- predict(mvt, interval = "prediction", conf_level = 0.95,
  newdata = d, type = "response")
pred_mvt <- pred_mvt %>% mutate(x = d$x, y = d$y, observed = d$cover,
  year = d$year, hold_out = d$hold_out)

pred_mvn <- predict(mvn, interval = "prediction", conf_level = 0.95,
  newdata = d, type = "response")
pred_mvn <- pred_mvn %>% mutate(x = d$x, y = d$y, observed = d$cover,
  year = d$year, hold_out = d$hold_out)

combined <- bind_rows(
  mutate(pred_mvt, model = "mvt"),
  mutate(pred_mvn, model = "mvn"))
```

Coverage for all the data:

```{r}
coverage <- dplyr::filter(combined) %>%
  group_by(model, year) %>%
  mutate(contained = observed < conf_high & observed > conf_low) %>%
  ungroup() %>%
  group_by(model) %>%
  summarize(coverage = mean(contained)) %>%
  print()
```

Coverage for just the held-out data:

```{r}
coverage <- dplyr::filter(combined, hold_out) %>%
  group_by(model, year) %>%
  mutate(contained = observed < conf_high & observed > conf_low) %>%
  ungroup() %>%
  group_by(model) %>%
  summarize(coverage = mean(contained)) %>%
  print()
```
